# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TripDestinationInput {
  /** @minLength 1 */
  destinationName: string;
  /** @minimum 1 */
  stayDays: number;
  /** @minimum 0 */
  roamingRadiusKm: number;
  notes?: string;
}

export interface TripInput {
  /** @minLength 1 */
  startLocation: string;
  startDate: string;
  endDate: string;
  /** @minimum 1 */
  passengerCount: number;
  vehicleType: string;
  tripType: string;
  specialRequirements?: string;
  /** @minItems 1 */
  destinations: TripDestinationInput[];
}

export type TripDestination = TripDestinationInput & {
  id: string;
  sequenceOrder: number;
};

export interface Trip {
  id: string;
  startLocation: string;
  startDate: string;
  endDate: string;
  passengerCount: number;
  vehicleType: string;
  tripType: string;
  specialRequirements?: string;
  destinations: TripDestination[];
  status: string;
  totalRouteDistanceKm?: number;
  estimatedDriveTime?: string;
  createdAt: string;
}

export interface RouteLeg {
  from: string;
  to: string;
  distanceKm: number;
  drivingTime: string;
}

export interface RouteSummary {
  tripId: string;
  totalDistanceKm: number;
  totalDrivingTime: string;
  legs: RouteLeg[];
}

export interface OperatorVehicle {
  type: string;
  label: string;
  seats: number;
  ac: boolean;
}

export interface Operator {
  id: string;
  businessName: string;
  operatorType: string;
  description: string;
  city: string;
  state: string;
  yearsExperience: number;
  verificationStatus: string;
  rating: number;
  totalReviews: number;
  maxTripDistanceKm: number;
  maxTripDurationDays: number;
  roamingCapabilityKm: number;
  vehicles: OperatorVehicle[];
  serviceAreas: string[];
  preferredDestinations: string[];
  tripTypes: string[];
  phone?: string;
  whatsapp?: string;
}

export interface OperatorMatch {
  operator: Operator;
  score: number;
  label: string;
  reasons: string[];
}

export interface EnquiryInput {
  operatorId: string;
  tripId: string;
  /** @minLength 1 */
  message: string;
}

export type Enquiry = EnquiryInput & {
  id: string;
  status: string;
  createdAt: string;
};

export interface DashboardSummary {
  activeTripCount: number;
  enquiryCount: number;
  savedOperatorCount: number;
  /** @nullable */
  nextTrip: string | null;
}

export type ListOperatorsParams = {
/**
 * @nullable
 */
tripId?: string | null;
};

export type MatchOperatorsParams = {
tripId: string;
};