# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  activeTripCount: number;
  enquiryCount: number;
  savedOperatorCount: number;
  /** @nullable */
  nextTrip: string | null;
}