# lib/api-zod/src/generated/types/enquiry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */
import type { EnquiryInput } from './enquiryInput';

export type Enquiry = EnquiryInput & {
  id: string;
  status: string;
  createdAt: string;
};