# lib/api-zod/src/generated/types/operator.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */
import type { OperatorVehicle } from './operatorVehicle';

export interface Operator {
  id: string;
  businessName: string;
  operatorType: string;
  description: string;
  city: string;
  state: string;
  yearsExperience: number;
  verificationStatus: string;
  rating: number;
  totalReviews: number;
  maxTripDistanceKm: number;
  maxTripDurationDays: number;
  roamingCapabilityKm: number;
  vehicles: OperatorVehicle[];
  serviceAreas: string[];
  preferredDestinations: string[];
  tripTypes: string[];
  phone?: string;
  whatsapp?: string;
}