# lib/api-zod/src/generated/types/operatorMatch.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */
import type { Operator } from './operator';

export interface OperatorMatch {
  operator: Operator;
  score: number;
  label: string;
  reasons: string[];
}