# lib/api-zod/src/generated/types/operatorVehicle.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface OperatorVehicle {
  type: string;
  label: string;
  seats: number;
  ac: boolean;
}