# lib/api-zod/src/generated/types/routeLeg.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface RouteLeg {
  from: string;
  to: string;
  distanceKm: number;
  drivingTime: string;
}