# lib/api-zod/src/generated/types/routeSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */
import type { RouteLeg } from './routeLeg';

export interface RouteSummary {
  tripId: string;
  totalDistanceKm: number;
  totalDrivingTime: string;
  legs: RouteLeg[];
}