# lib/api-zod/src/generated/types/trip.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */
import type { TripDestination } from './tripDestination';

export interface Trip {
  id: string;
  startLocation: string;
  startDate: string;
  endDate: string;
  passengerCount: number;
  vehicleType: string;
  tripType: string;
  specialRequirements?: string;
  destinations: TripDestination[];
  status: string;
  totalRouteDistanceKm?: number;
  estimatedDriveTime?: string;
  createdAt: string;
}