# lib/api-zod/src/generated/types/tripDestination.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */
import type { TripDestinationInput } from './tripDestinationInput';

export type TripDestination = TripDestinationInput & {
  id: string;
  sequenceOrder: number;
};