# lib/api-zod/src/generated/types/tripDestinationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * TripQuote travel planning and operator discovery API
 * OpenAPI spec version: 0.1.0
 */

export interface TripDestinationInput {
  /** @minLength 1 */
  destinationName: string;
  /** @minimum 1 */
  stayDays: number;
  /** @minimum 0 */
  roamingRadiusKm: number;
  notes?: string;
}